requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
numpy==1.26.4
torch==2.5.1
torchvision==0.20.1
ultralytics==8.3.57
tensorflow==2.18.0
opencv-python-headless==4.10.0.84
timm==1.0.12
gunicorn==23.0.0
